% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htab.R
\name{scale_terrain}
\alias{scale_terrain}
\title{function to scale values}
\usage{
scale_terrain(x, n)
}
\arguments{
\item{x}{vector to be scaled}

\item{n}{number of colors}
}
\description{
function to scale values
}

